[org.notepad_plus_plus.Notepad-plus-plus] Use pre-build binaries
This switches the installer from using the npp installer to the
pre-build binaries. Files are extracted to /app/extra/notepad++ and
then a copy is made to /var/data/notepad++. The notepad++ directory
has to be writable so the executable can't be run from /app/extra.

Since this removes the installer it should speed up the install
processes and should make it harder to fail.