Add com.blizzard.BattleNet.BaseApp
This commit as a BaseApp for Battle.net Games. In-order to use the BaseApp you
need to add it to your manifest:

    base: com.blizzard.BattleNet.BaseApp
    base-version: stable

and then inherit some of the extensions.

    inherit-extensions:
      - org.winepak.Platform.Compat32
      - org.winepak.Platform.Extension.corefonts

Then you can run the installer from your manifest:

    source /app/bin/battlenet-installer

The Battle.net installer also helps to install Battle.net games by supporting
the battlenet:// url scheme. eg.

    source /app/bin/battlenet-installer "battlenet://starcraft"

A full list of schemes include:

- battlenet://WoW         World of Warcraft
- battlenet://D3          Diablo 3
- battlenet://WTCG        Hearthstone
- battlenet://Pro         Overwatch
- battlenet://starcraft   Starcraft 2
- battlenet://heroes      Heroes of the Storm
- battlenet://SCR         Starcraft Remastered
- battlenet://DST2        Destiny 2