Move to Calendar Versioning
The concept of MAJOR.MINOR.MICRO versioning doesn't really make sense for
winepak. We don't actually control the application's version number and our
releases aren't synced with Wine. What's a MAJOR, MINOR, or MICRO update to a
manifest, what is the criteria? A major wine bump? A measurable improvement in
performance? A change to the actual application? We make updates and
improvements at random intervals during the year. Versioning by date makes more
sense and shows users how "fresh" a winepak-ed application is.